separate the function
